graphical vesion of Minesweeper
This is a version of Minesweeper. I write this vesion in VS 2012, and
used  graphic library named EasyX, which is free but not open source. I
simplely copy all of the project to this repository(some files are
ignored by GitHub defaultly). It's not until I completed this version
that I realized that I forget to insert a help function to instruct the
game's users how to play. It's a pity but also fun, I think, to dig up
potential operations.